Add files via upload
Please find my project - BuildComAsses for the test automation assessment requested. 
This is a runnable JAR file you will need to execute and the project will start immediately. 
The system should run test flows on Chrome and FireFox browser simultaneously and would submit and open a report file for each server at the end of the process.
The system works best with Chrome v. 61 and FireFox v. 55. however the Chrome version was not tested enough since the CAPTCHA interfere in the process and made it difficult. CAPTCHA page should be disabled on the server when creating an automated test system.